subscriber/requirements.txt
paho-mqtt==1.6.1
redis==5.0.1
